pyproject.toml
[tool.poetry]
name = "piano-project"
version = "0.1.0"
description = "\"\""
authors = ["Deeptanshu Sanghani <deeptanshusanghani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.31"
passlib = "^1.7.4"
google-cloud-storage = "^2.18.0"
python-dotenv = "^1.0.1"
google-cloud-firestore = "^2.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
